Refactor to separate ignore functionality.
